requirements.txt
seleniumbase
streamlit
streamlit-extras
numpy
attrs
pandas
plotly
pip
matplotlib
watchdog
scikit-learn
statsmodels